% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sms_functions.R
\name{average_ghi_sensors}
\alias{average_ghi_sensors}
\title{Average GHI Sensors}
\usage{
average_ghi_sensors(sms_flagged)
}
\arguments{
\item{sms_flagged}{An sms dataframe with tidy variables}
}
\value{
sms_flagged with an additional variable ghi_ave
}
\description{
Average redundant ghi sensors of onsite dataframe
}
\details{
If there are 3 or more sensors, conduct a straight average at each datapoint, ignoring missing data.
If there are only 2 sensors, create a reconstructed average, by applying the monthly bias between the two sensors to reconstruct any missing data points before averaging.
}
\examples{
\dontrun{
## To Prepare sms_flagged dataframe for megaframe compliation:
sms_flagged \%>\%
  filter_sms() \%>\%
  average_ghi_sensors() \%>\%
  average_data_to_hourly() \%>\%
  format_sms_vars_for_megaframe()
}
}
