% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/satellite_functions.R
\name{compile_megaframe}
\alias{compile_megaframe}
\title{Build a Megaframe!}
\usage{
compile_megaframe(sms, cpr_list, vaisala_list)
}
\arguments{
\item{sms}{Groundwork data that has been filtered, sensors have been averaged, and minute data has been averaged to hourly}

\item{cpr_list}{Tidy CPR data}

\item{vaisala_list}{Tidy Vaisala data}
}
\value{
A single dataframe that includes all of the sms, cpr, and vaisala timeseries data in a tidy format
}
\description{
Compiles tidy sms, cpr, and vaisala data frames into one MEGAFRAME!
}
\details{
Megaframe format includes hourly variables: datetime, year, month, ghi, dni, dhi, temp, ws, model
}
\examples{
\dontrun{
## Prepare the inputs
sms_flagged <- read_groundwork('groundwork.csv')
sms <- prepare_sms_flagged_for_megaframe(sms_flagged)
cpr_list <- read_cpr(c('path/to/cpr33.csv', 'path/to/cpr34.csv'))
vaisala_list <- read_vaisala('path/to/vaisala213.csv')

## Create the Megaframe!
mf <- compile_megaframe(sms, cpr_list, vaisala_list)
head(mf)
}
}
