% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{count_flags}
\alias{count_flags}
\title{Count Flags}
\usage{
count_flags(sms_flagged)
}
\arguments{
\item{sms_flagged}{A dataframe of onsite measurements and flags with tidy variables}
}
\value{
A dataframe with variables month, sensor, flag_type, n, perc where n is the
count of flags and perc is the percent of datapoints for that month/sensor
}
\description{
Count the number of flags per month for the onsite measurements
}
\examples{
\dontrun{
flag_count <- count_flags(sms_flagged)

plot_flag_count_bar_chart(flag_count)
}
}
