% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sms_functions.R
\name{filter_sms}
\alias{filter_sms}
\title{Filter SMS data}
\usage{
filter_sms(
  sms_flagged,
  ghi_cutoff_min = 5,
  ws_cutoff_min = 0.25,
  flagged_data_is_filtered = TRUE
)
}
\arguments{
\item{sms_flagged}{Dataframe of sms data with "tidy" variables}

\item{ghi_cutoff_min}{Measured GHI values below this cutoff will be filtered to zero because of poor sensor performance at low values (W/m2)}

\item{ws_cutoff_min}{Measured wind speed values below this cutoff will be filtered to zero (m/s)}

\item{flagged_data_is_filtered}{Whether or not we should filter out the flagged data (Recommended)}
}
\value{
SMS Dataframe with zero at all points below the measurement cutoff minimum and NA for all points with flag filters applied
}
\description{
Apply measurement cutoff filters and flag filters
}
\examples{
\dontrun{
## Get flagged, tidy sms data
sms_flagged <- read_groundwork(groundwork_file_list)

## Apply filters
sms_filtered <- filter_sms(sms_flagged)
no_flag_filters <- filter_sms(sms_flagged, flagged_data_is_filtered = FALSE)
change_cutoff_filters <- filter_sms(sms_flagged, ghi_cutoff_min = 10, ws_cutoff_min = 0)
}
}
