% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sms_functions.R
\name{format_sms_vars_for_megaframe}
\alias{format_sms_vars_for_megaframe}
\title{Format SMS Variables for Megaframe}
\usage{
format_sms_vars_for_megaframe(sms_flagged)
}
\arguments{
\item{sms_flagged}{SMS dataframe with "tidy" variables, including ghi_ave}
}
\value{
SMS dataframe with variables ready to integrate with Megaframe format
}
\description{
Selects/creates the variables needed for megaframe, using ghi_ave for ghi
}
\details{
Megaframe format variables: datetime, year, month, ghi, dni, dhi, temp, ws, model
}
\examples{
\dontrun{
## To Prepare sms_flagged dataframe for megaframe compliation:
sms_flagged \%>\%
  filter_sms() \%>\%
  average_ghi_sensors() \%>\%
  average_data_to_hourly() \%>\%
  format_sms_vars_for_megaframe()
}
}
