% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_functions.R
\name{get_longterm_adjustment}
\alias{get_longterm_adjustment}
\title{Adjust Longterm Resource for all timesteps and sensors}
\usage{
get_longterm_adjustment(mcp_list, longterm_df, model_coefficient_is_0 = FALSE)
}
\arguments{
\item{mcp_list}{The results of run_mcp_for_all_timesteps_and_sensors()}

\item{longterm_df}{The results of build_longterm_resource_with_all_sensors}

\item{model_coefficient_is_0}{Whether or not the models in mcp_list were forced to have a zero intercept coefficient}
}
\value{
dataframe with longterm adjusted resource for each model and sensor
}
\description{
Adjust Longterm Resource for all timesteps and sensors
}
