% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_functions.R
\name{get_longterm_metrics}
\alias{get_longterm_metrics}
\title{Get Longterm Metrics}
\usage{
get_longterm_metrics(megaframe)
}
\arguments{
\item{megaframe}{The dataframe containing the longterm models and resources}
}
\value{
A dataframe with the variabels

por: period of record of all models and sensors

sms_por: period of record for the onsite measurements

iav: Inter-annual variability for each model and resource
}
\description{
Get the longterm metrics for all the models and sensors in a megaframe
}
