% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_functions.R
\name{get_longterm_resource}
\alias{get_longterm_resource}
\title{Get Longterm Resource}
\usage{
get_longterm_resource(megaframe)
}
\arguments{
\item{megaframe}{The dataframe containing the longterm satellite models and resources}
}
\value{
Returns the longterm average for wind speed or temperature and the average annual sum in kWh/m2 for ghi
}
\description{
Get the longterm resource for all the models and sensors in a megaframe
}
