% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_functions.R
\name{get_mcp_results}
\alias{get_mcp_results}
\title{Calculate the results of the MCP for all timesteps and sensors}
\usage{
get_mcp_results(mcp_list, megaframe, model_coefficient_is_0 = FALSE)
}
\arguments{
\item{mcp_list}{A named list of model fits to onsite data; The output of run_mcp_for_all_timesteps_and_sensors()}

\item{megaframe}{A megaframe with at least one satellite model and the sms model}

\item{model_coefficient_is_0}{Whether or not the models in mcp_list forced the intercept coefficient to zero}
}
\value{
A dataframe that contains all of the result metrics for the mcp runs in mcp_list
}
\description{
Build the dataframe that is saved in MCP$results
Gets longterm metrics from the satellite models, gets performance metrics of the model fits, and calculates longterm adjustments
}
\examples{
\dontrun{
my_mcp <- run_mcp(megaframe)
get_mcp_restults(my_mcp, megaframe)
}
}
