% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_functions.R
\name{get_timestep_results_table}
\alias{get_timestep_results_table}
\title{Build LTRA Table}
\usage{
get_timestep_results_table(mcp_results_df, model_timestep = "daily")
}
\arguments{
\item{mcp_results_df}{The results of an MCP}

\item{model_timestep}{The timestep you want to include in the table}
}
\value{
A df the is "wider" with metrics more easily identifiable
}
\description{
Modifies the results of an MCP to fit better with a table for a specific timestep
}
