% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sms_functions.R
\name{import_groundwork_files}
\alias{import_groundwork_files}
\title{Import Groundwork Onsite Data}
\usage{
import_groundwork_files(sms_files)
}
\arguments{
\item{sms_files}{A list of Groundwork CSVs to import}
}
\value{
A dataframe containing all of the raw onsite data
}
\description{
Import Groundwork QC Data CSVs without doing any variable tidying
}
\examples{
\dontrun{
## Get the files to import
sms_files <- paste0('data/Groundwork/', dir(path = "data/Groundwork/", pattern = "*.csv"))

## Import the files
groundwork_df <- import_groundwork_files(sms_files)

## Tidy the Dataframe
tidy_sms <- tidy_groundwork(groundwork_df)

## Flag the Dataframe
sms_flagged <- flag_sms(tidy_sms)

}
}
