% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_app.R
\name{launch_app}
\alias{launch_app}
\title{Launch App}
\usage{
launch_app()
}
\description{
Launch a graphical user interface that allows you to upload measured data and longterm satellite models and run an MCP.  Has the ability to generate reports and export adjusted TMYs.
}
\examples{
\dontrun{
launch_app()
}
}
