% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCP_class.R
\docType{class}
\name{MCP-class}
\alias{MCP-class}
\alias{MCP}
\title{MCP Class}
\arguments{
\item{megaframe}{A dataframe that contains all of the measured and modeled data in a specific, tidy format}

\item{minimum_required_hours_for_daily_inclusion}{The number of hourly data points required to include a day in the analysis}

\item{model_coefficient_is_0}{Should the linear regression force the intercept to zero?}
}
\description{
This class contains all of the inputs and results related to an MCP.
An "MCP" is an acronym for Measure Correlate Predict, which is a common approach to correct the bias between
measured data and longterm, modeled data.  The methodology is to fit a linear regression of the concurrent
measured and modeled data and then use that linear fit to predict future data.
}
\section{Fields}{

\describe{
\item{\code{results}}{A dataframe containing the results of an MCP run}

\item{\code{daily_results_table}}{A dataframe of the results formatted to show daily timestep results more clearly}

\item{\code{hourly_results_table}}{A dataframe of the results formatted to show hourly timestep results more clearly}
}}

\section{Methods}{

\describe{
\item{\code{get_daily_results_table()}}{Get Results of daily-timestep MCP runs, formatted in an easy-to-read table format.}

\item{\code{get_hourly_results_table()}}{Get Results of hourly-timestep MCP runs, formatted in an easy-to-read table format.}

\item{\code{get_results()}}{Get Results of all MCP runs}

\item{\code{initialize(
  megaframe = megaframe,
  minimum_required_hours_for_daily_inclusion = 6,
  model_coefficient_is_0 = FALSE
)}}{Creates an instance of the class.  Call with MCP().}

\item{\code{run()}}{Run an MCP for all timesteps (daily and hourly) and sensors (ghi, ws, temp).}
}}

\examples{
\dontrun{
my_mcp <- MCP(megaframe)
my_mcp$run()
my_mcp$get_results()
}
}
