% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_campaign_ghi_longterm_adjustments}
\alias{plot_campaign_ghi_longterm_adjustments}
\title{Plot Campaign GHI Longterm Adjustments}
\usage{
plot_campaign_ghi_longterm_adjustments(campaign_ghi_results)
}
\arguments{
\item{campaign_ghi_results}{output of run_sms_campaign()}
}
\value{
ggplot lineplot that shows change in GHI Longterm Adjustment for each model after each month
of the SMS campaign
}
\description{
Plot Campaign GHI Longterm Adjustments
}
\examples{
\dontrun{
campaign_results <- run_sms_campaign(megaframe)
plot_campaign_ghi_longterm_adjustments(campaign_results)
}
}
