% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_campaign_rom_uncertainty}
\alias{plot_campaign_rom_uncertainty}
\title{Plot Campaign GHI Uncertainty}
\usage{
plot_campaign_rom_uncertainty(campaign_ghi_results)
}
\arguments{
\item{campaign_ghi_results}{output of run_sms_campaign()}
}
\value{
ggplot lineplot that shows change in Representativeness of Modeling Period Uncertainty after each month
of the SMS campaign
}
\description{
Plot Campaign GHI Uncertainty
}
\examples{
\dontrun{
campaign_results <- run_sms_campaign(megaframe)
plot_campaign_rom_uncertainty(campaign_results)
}
}
