% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_flag_count_bar_chart}
\alias{plot_flag_count_bar_chart}
\title{Plot Flag Count}
\usage{
plot_flag_count_bar_chart(flag_count)
}
\arguments{
\item{flag_count}{Dataframe, the output of count_flags()}
}
\value{
ggplot bar chart showing percent breakdown of flag_types by month and sensor
}
\description{
Plot Flag Count
}
\examples{
\dontrun{
flag_count <- count_flags(sms_flagged)

plot_flag_count_bar_chart(flag_count)
}
}
