% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sms_functions.R
\name{prepare_sms_flagged_for_megaframe}
\alias{prepare_sms_flagged_for_megaframe}
\title{Prepare Flagged SMS Dataframe for Megaframe Compilation}
\usage{
prepare_sms_flagged_for_megaframe(
  sms_flagged,
  ghi_cutoff_min = 5,
  ws_cutoff_min = 0.25,
  flagged_data_is_filtered = TRUE
)
}
\arguments{
\item{sms_flagged}{Tidy SMS dataframe}

\item{ghi_cutoff_min}{Measured GHI values below this cutoff will be filtered to zero because of poor sensor performance at low values (W/m2)}

\item{ws_cutoff_min}{Measured wind speed values below this cutoff will be filtered to zero (m/s)}

\item{flagged_data_is_filtered}{Whether or not we should filter out the flagged data (Recommended)}
}
\value{
dataframe of onsite data formatted for megaframe inclusion
}
\description{
Converts a tidy sms (solar met station) dataframe to megaframe format
}
\details{
megaframe format consists of distinct, hourly variabls:
datetime, year, month, model, ghi, dni, dhi, temp, ws
}
\examples{
\dontrun{
## Get sms_flagged dataframe (onsite data with "tidy" variables)
sms_files <- paste0('data/Groundwork/', dir(path = "data/Groundwork/", pattern = "*.csv"))
sms_flagged <- read_groundwork(sms_files)

## Prepare for Megaframe
sms <- prepare_sms_flagged_for_megaframe(sms_flagged)
}
}
