% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/satellite_functions.R
\name{read_cpr}
\alias{read_cpr}
\title{Read CPR CSVs into a list of tidy dataframes}
\usage{
read_cpr(cpr_files)
}
\arguments{
\item{cpr_files}{A list of strings that provide a path to each CPR Timeseries CSV.}
}
\value{
A list of tidy data frames (technically tibbles), named after the CPR model version
}
\description{
Reads in a list of Clean Power Research Time Series Files and returns a list of data frames named after each model version
and formatted to be integrated with a Megaframe.
}
\details{
Megaframe format variables: datetime, year, month, ghi, dni, dhi, temp, ws, model
}
\examples{
\dontrun{
cpr_list <- read_cpr(c('path/to/cpr32.csv', 'path/to/cpr33.csv', 'path/to/cpr34.csv'))
head(cpr_list$cpr3.3)

cpr_list <- read_cpr('cpr34.csv')
head(cpr_list$cpr3.4)
}
}
