% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sms_functions.R
\name{read_groundwork}
\alias{read_groundwork}
\title{Import and Tidy Groundwork Onsite Data}
\usage{
read_groundwork(sms_files)
}
\arguments{
\item{sms_files}{A list of Groundwork CSVs to import}
}
\value{
A dataframe containing all of the raw onsite data and flags that can be used to apply filters
}
\description{
Import Groundwork QC Data CSVs into a single dataframe with clean variable names
}
\details{
Variable naming convention: datetime, month, year, ghi1, ghi2, ghi3, ws, temp, ghi1_flag, ghi2_flag, ghi3_flag, ws_flag, temp_flag
}
\examples{
\dontrun{
## Read in a list of files
sms_flagged <- read_groundwork(c('SiteDecember2019QCData.csv', 'SiteJanuary2020QCData.csv'))
head(sms_flagged)

## Read in a single file
sms_flagged <- read_groundwork('Site2019QCData.csv')

## Read in all files in a folder that end in csv
sms_files <- paste0('data/Groundwork/', dir(path = "data/Groundwork/", pattern = "*.csv"))
sms_flagged <- read_groundwork(sms_files)
summary(sms_flagged)
}
}
