% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/satellite_functions.R
\name{read_vaisala}
\alias{read_vaisala}
\title{Read Vaisala CSVs into a list of tidy dataframes}
\usage{
read_vaisala(vaisala_files)
}
\arguments{
\item{vaisala_files}{A list of strings that provide a path to each Vaisala Timeseries CSV.}
}
\value{
A list of tidy data frames (technically tibbles), named after the Vaisala model version
}
\description{
Reads in a list of Vaisala Time Series Files and returns a list of data frames named after each model version
and formatted to be integrated with a Megaframe.
}
\details{
Megaframe format variables: datetime, year, month, ghi, dni, dhi, temp, ws, model
}
\examples{
\dontrun{
vaisala_list <- read_vaisala(c('path/to/vaisala123.csv', 'path/to/vaisala213.csv'))
head(vaisala_list$vaisala2.1.3)

vaisala_list <- read_vaisala('vaisala123.csv')
head(vaisala_list$vaisala1.2.3)
}
}
