% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_functions.R
\name{run_mcp}
\alias{run_mcp}
\title{Run MCP for all Timesteps and Sensors}
\usage{
run_mcp(
  megaframe,
  minimum_required_hours_for_daily_inclusion = 6,
  model_coefficient_is_0 = FALSE
)
}
\arguments{
\item{megaframe}{A megaframe with at least one satellite model and the sms model}

\item{minimum_required_hours_for_daily_inclusion}{The number of data points required for a day to be included in the regression}

\item{model_coefficient_is_0}{Whether or not the linear regression should be forced to have a zero intercept}
}
\value{
A list of model fits to the onsite data, organized by timestep->sensor->model
}
\description{
The function called by the run() method of the MCP class.
}
\examples{
\dontrun{
my_mcp <- run_mcp(megaframe)
get_mcp_restults(my_mcp)

## Run MCP for a specific sensor
mf <- filter(megaframe, sensor == 'ghi')
my_mcp <- run_mcp(mf)
get_mcp_results(my_mcp, mf)
}
}
