% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_functions.R
\name{run_mcp_for_single_timestep}
\alias{run_mcp_for_single_timestep}
\title{Run MCP for all Sensors at one timestep}
\usage{
run_mcp_for_single_timestep(
  megaframe,
  daily = TRUE,
  minimum_required_hours_for_daily_inclusion = 6,
  model_coefficient_is_0 = FALSE
)
}
\arguments{
\item{megaframe}{A megaframe with at least one satellite model and the sms model}

\item{daily}{Logical; daily timestep if true, hourly if false}

\item{minimum_required_hours_for_daily_inclusion}{The number of data points required for a day to be included in the regression}

\item{model_coefficient_is_0}{Whether or not the linear regression should be forced to have a zero intercept}
}
\value{
A list of model fits to the onsite data, organized by timestep->sensor->model
}
\description{
run_mcp() but for either daily or hourly
}
\examples{
\dontrun{
daily_mcp <- run_mcp_for_single_timestep(megaframe, daily = TRUE)
get_mcp_results(daily_mcp, megaframe)

hourly_mcp <- run_mcp_for_single_timestep(megaframe, daily = FALSE)
get_mcp_results(hourly_mcp, megaframe)

## Run MCP for a specific sensor and timestep
mf <- select(megaframe, -temp, -ws)
my_mcp <- run_mcp_for_single_timestep(mf)
get_mcp_results(my_mcp, mf)
}
}
