% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{scale_tmy}
\alias{scale_tmy}
\title{Scale TMY}
\usage{
scale_tmy(mcp_results, tmy, model_name, model_timestep, sensor_list)
}
\arguments{
\item{mcp_results}{Results of an MCP}

\item{tmy}{a Tidy tmy}

\item{model_name}{the model in mcp_results to be used as the long-term resource adjustment}

\item{model_timestep}{the timestep in mcp_results to be used as the long-term resource adjustment}

\item{sensor_list}{the sensors of the tmy to be scaled}
}
\value{
The original tmy with the sensors scaled to the long-term adjusted values.
If 'ghi' is one of the sensors, dhi and dni will be scaled by the same ratio as ghi.
}
\description{
Adjust the tmy to reflect the longterm resource adjustment calculated by the MCP
}
\examples{
\dontrun{
my_mcp <- MCP(megaframe)
my_mcp$run()

tmy <- reac_csv('tmy.csv') #columns = datetime, ghi, dni, dhi, temp, ws

adjusted_tmy <- scale_tmy(my_mcp$get_results(), tmy, 'cpr3.4', 'daily', c('ghi', 'temp'))
}
}
