% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/satellite_functions.R
\name{tidy_vaisala}
\alias{tidy_vaisala}
\title{Tidy Vaisala Data Frame}
\usage{
tidy_vaisala(vaisala_df)
}
\arguments{
\item{vaisala_df}{A dataframe from importing a Vaisala timeseries}
}
\value{
A dataframe in the Megaframe format
}
\description{
Converts an imported Vaisala data frame to one that is in the Megaframe format
}
\details{
Megaframe format variables: datetime, year, month, ghi, dni, dhi, temp, ws, model
}
\examples{
\dontrun{
vaisala_df <- read_csv('vaisala_file.csv', skip = 1)
vaisala <- tidy_vaisala(vaisala_df)
summary(vaisala)
}
}
